Resolve minor issue creating subscriptions with descriptions
When creating a subscription with a description, the description
was not being saved (and as subscriptions can't be patched there
was no way of adding this field via the API).

This commit updates:
 * silver/tests/spec/test_subscription.py
   * adding two functions to check that subscriptions created with a reference and description respectivey, record this information
 * silver/api/serializers.py
   * fix so JSONSerializerField.to_internal_value that it passes hooks/pre-commit (the script was complaining about 'silver/api/serializers.py:93:17: W503 line break before binary operator', and I've moved the 'and' onto the previous line to resolve this)
   * added 'description' to the list of files in the SubscriptionSerializer so that you can create subscriptions with descriptions

The test to check the reference when adding a subscription is probably
unnecessary (as it worked before), but seems to do no harm.